backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
groq
python-dotenv
sqlalchemy
psycopg2-binary
